refactor: extract workspace filter predicates from workbench listing
Keeps workbench_listing.dart under the 500-line max-lines ratchet.
